Projeto ITP

+++ Container/Materias/Periodo 1/Introdução às Técnicas de Programação/Atividades/ProjetoU3/src/ColorMap.cpp
@@ -1,6 +1,6 @@
 #include "ColorMap.hpp"
 
-colors::ColorMap::ColorMap(unsigned short r, unsigned g, unsigned short b, double height) 
+colors::ColorMap::ColorMap(int r, int g, int b, double height) 
 : colors::Color(r, g, b), _height(height)
 {
